move RecordCursor into Session module
